moving qiskit zvode
