+++ src/open_pulse/CMakeLists.txt
@@ -9,6 +9,6 @@ include(cython_utils)
 # shared library.
 string(REPLACE " -static " "" CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}")
 
-set(CYTHON_INSTALL_DIR "qiskit/providers/aer/pulse/pulse0/cy/")
+set(CYTHON_INSTALL_DIR "qiskit/providers/aer/pulse/de_solvers/")
 add_cython_module(numeric_integrator_wrapper numeric_integrator.cpp)
 add_cython_module(test_python_to_cpp)
